Добавить метод проверки в MainSQLPage checkLogIn()
