+++ src/test/java/com/way2automation/tests/BaseTest.java
@@ -7,14 +7,13 @@ import org.openqa.selenium.chrome.ChromeOptions;
 import org.openqa.selenium.remote.CapabilityType;
 import org.testng.annotations.*;
 
-import java.io.FileNotFoundException;
 import java.util.concurrent.TimeUnit;
 
 public class BaseTest {
     protected static WebDriver driver;
 
     @BeforeMethod
-    public void setUp() throws FileNotFoundException {
+    public void setUp() {
         ChromeOptions options = new ChromeOptions();
         options.setCapability(CapabilityType.PAGE_LOAD_STRATEGY, "eager");
         driver = new ChromeDriver(options);
